requirements.txt

fastapi>=0.95
uvicorn[standard]>=0.20
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7
jinja2>=3.0
python-multipart>=0.0.6
argon2-cffi>=21.3.0